simple-agent/pyproject.toml
[project]
name = "simple-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-google-genai>=2.1.9",
    "langchain-mcp-adapters>=0.1.4",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
    "streamlit>=1.48.1",
]
